RE organizacion del codigo
